+++ automation/src/test/scala/org/broadinstitute/dsde/firecloud/component/Clickable.scala
@@ -25,19 +25,15 @@ trait Clickable extends LazyLogging { this: Component =>
     try {
       click on query
     } catch {
-      // on rare occasion, stale exception happens on click
-      case e: StaleElementReferenceException => click on query
       case e: WebDriverException =>
         logger.warn("doClick: " + e.getMessage)
-        // make an attempt to recover when this exact error occurred
+        // attempt to recover when this exact error occurred
         if (e.getMessage.contains("Other element would receive the click")) {
           logger.warn("retrying \"click on query\" after sleep 5 seconds")
           Thread.sleep(5000)
           click on query
         }
-    } finally {
-      Thread sleep 100
-      await notVisible cssSelector("[data-test-id=spinner]")
     }
   }
+
 }
