+++ automation/src/test/scala/org/broadinstitute/dsde/firecloud/component/TabBar.scala
@@ -7,6 +7,7 @@ case class TabBar(private val queryString: QueryString = TestId("tabs"))(implici
   def goToTab(name: String): Unit = {
     awaitReady()
     Link(s"$name-tab" inside this).doClick()
+    await notVisible cssSelector("[data-test-id=spinner]")
   }
 
 }
