+++ automation/src/test/scala/org/broadinstitute/dsde/firecloud/test/analysis/MethodRedactedSpec.scala
@@ -22,9 +22,10 @@ class MethodRedactedSpec extends FreeSpec with Matchers with WebBrowserSpec with
       withCleanBillingProject(user) { billingProject =>
         withWorkspace(billingProject, "MethodRedactedSpec_choose_new_snapshot") { workspaceName =>
           withMethod("MethodRedactedSpec_choose_new_snapshot", MethodData.SimpleMethod, 2) { methodName =>
-            api.methodConfigurations.createMethodConfigInWorkspace(billingProject, workspaceName, MethodData.SimpleMethod.copy(methodName = methodName),
+            val method = MethodData.SimpleMethod.copy(methodName = methodName)
+            api.methodConfigurations.createMethodConfigInWorkspace(billingProject, workspaceName, method,
               SimpleMethodConfig.configNamespace, methodConfigName, 1, SimpleMethodConfig.inputs, SimpleMethodConfig.outputs, "participant")
-            api.methods.redact(MethodData.SimpleMethod.copy(methodName = methodName))
+            api.methods.redact(method)
 
             withSignIn(user) { _ =>
               val methodConfigDetailsPage = new WorkspaceMethodConfigDetailsPage(billingProject, workspaceName, SimpleMethodConfig.configNamespace, methodConfigName).open
@@ -50,9 +51,10 @@ class MethodRedactedSpec extends FreeSpec with Matchers with WebBrowserSpec with
       withCleanBillingProject(user) { billingProject =>
         withWorkspace(billingProject, "MethodRedactedSpec_delete") { workspaceName =>
           withMethod("MethodRedactedSpec_delete", MethodData.SimpleMethod, cleanUp = false) { methodName =>
-            api.methodConfigurations.createMethodConfigInWorkspace(billingProject, workspaceName, MethodData.SimpleMethod.copy(methodName = methodName),
+            val method = MethodData.SimpleMethod.copy(methodName = methodName)
+            api.methodConfigurations.createMethodConfigInWorkspace(billingProject, workspaceName, method,
               SimpleMethodConfig.configNamespace, methodConfigName, 1, SimpleMethodConfig.inputs, SimpleMethodConfig.outputs, "participant")
-            api.methods.redact(MethodData.SimpleMethod.copy(methodName = methodName))
+            api.methods.redact(method)
 
             withSignIn(user) { _ =>
               // launch analysis error:
@@ -77,88 +79,6 @@ class MethodRedactedSpec extends FreeSpec with Matchers with WebBrowserSpec with
               val list = await ready new WorkspaceMethodConfigListPage(billingProject, workspaceName)
               list.hasConfig(methodConfigName) shouldBe false
             }
-          }
-        }
-      }
-    }
-  }
-
-  "owner can abort a launched submission" in withWebDriver { implicit driver =>
-    val user = Config.Users.owner
-    implicit val authToken: AuthToken = user.makeAuthToken()
-    withCleanBillingProject(user) { billingProject =>
-      withWorkspace(billingProject, "MethodRedactedSpec_abort_submission") { workspaceName =>
-        api.workspaces.waitForBucketReadAccess(billingProject, workspaceName)
-
-        val shouldUseCallCaching = false
-        api.importMetaData(billingProject, workspaceName, "entities", TestData.SingleParticipant.participantEntity)
-
-        withMethod("MethodRedactedSpec_abort", MethodData.SimpleMethod) { methodName =>
-          api.methodConfigurations.createMethodConfigInWorkspace(billingProject, workspaceName,
-            MethodData.SimpleMethod.copy(methodName = methodName), SimpleMethodConfig.configNamespace, methodName, 1,
-            SimpleMethodConfig.inputs, SimpleMethodConfig.outputs, MethodData.SimpleMethod.rootEntityType)
-
-          withSignIn(user) { _ =>
-            val methodConfigDetailsPage = new WorkspaceMethodConfigDetailsPage(billingProject, workspaceName, SimpleMethodConfig.configNamespace, methodName).open
-            val submissionDetailsPage = methodConfigDetailsPage.launchAnalysis(MethodData.SimpleMethod.rootEntityType, TestData.SingleParticipant.entityId, "", shouldUseCallCaching)
-
-            submissionDetailsPage.abortSubmission()
-            submissionDetailsPage.waitUntilSubmissionCompletes()
-            submissionDetailsPage.getSubmissionStatus shouldBe submissionDetailsPage.ABORTED_STATUS
-          }
-        }
-      }
-    }
-  }
-
-  "import method config" - {
-    "copy from a workspace" in withWebDriver { implicit driver =>
-      val user = UserPool.chooseProjectOwner
-      implicit val authToken: AuthToken = user.makeAuthToken()
-      withCleanBillingProject(user) { billingProject =>
-        withWorkspace(billingProject, "Test_copy_method_config_from_workspace_src") { sourceWorkspaceName =>
-          withWorkspace(billingProject, "Test_copy_method_config_from_workspace_dest") { destWorkspaceName =>
-            val method = MethodData.SimpleMethod
-            withMethod("MethodRedactedSpec_import_from_workspace", method, 1) { methodName =>
-              api.methodConfigurations.createMethodConfigInWorkspace(billingProject, sourceWorkspaceName,
-                method.copy(methodName = methodName), method.methodNamespace, method.methodName, 1, Map.empty, Map.empty, method.rootEntityType)
-
-              withSignIn(user) { listPage =>
-                val methodConfigTab = listPage.enterWorkspace(billingProject, destWorkspaceName).goToMethodConfigTab()
-
-                val methodConfigDetailsPage = methodConfigTab.copyMethodConfigFromWorkspace(
-                  billingProject, sourceWorkspaceName, method.methodNamespace, method.methodName)
-
-                methodConfigDetailsPage.isLoaded shouldBe true
-                methodConfigDetailsPage.methodConfigName shouldBe method.methodName
-
-                // launch modal shows no default entities
-                val launchModal = methodConfigDetailsPage.openLaunchAnalysisModal()
-                launchModal.verifyNoRowsMessage() shouldBe true
-                launchModal.xOut()
-              }
-            }
-          }
-        }
-      }
-    }
-
-    "import from method repo" in withWebDriver { implicit driver =>
-      val user = UserPool.chooseProjectOwner
-      implicit val authToken: AuthToken = user.makeAuthToken()
-      withCleanBillingProject(user) { billingProject =>
-        withWorkspace(billingProject, "MethodRedactedSpec_import_from_methodrepo") { workspaceName =>
-          withSignIn(user) { workspaceListPage =>
-            val methodConfigPage = workspaceListPage.enterWorkspace(billingProject, workspaceName).goToMethodConfigTab()
-
-            val methodConfigDetailsPage = methodConfigPage.importMethodConfigFromRepo(
-              MethodData.SimpleMethod.methodNamespace,
-              MethodData.SimpleMethod.methodName,
-              MethodData.SimpleMethod.snapshotId,
-              SimpleMethodConfig.configName)
-
-            methodConfigDetailsPage.isLoaded shouldBe true
-            methodConfigDetailsPage.editMethodConfig(inputs = Some(SimpleMethodConfig.inputs))
           }
         }
       }
